Move the useragent and javascript eval helper functions into their own service.
So that they may be easily used by BaseProviders and BaseResolvers as needed.
