+++ src/scrapers/providers/BaseProvider.js
@@ -4,7 +4,8 @@ const Promise = require('bluebird');
 const RequestPromise = require('request-promise');
 const resolve = require('../resolvers/resolve');
 const logger = require('../../utils/logger');
-const randomUseragent = require('random-useragent');
+const javascriptEval = require('../../service/javascriptEval');
+const useragent = require('../../service/useragent');
 
 function _implementMe(functionName) {
     throw new Error(`Must implement ${functionName}()`);
@@ -19,7 +20,7 @@ function _implementMe(functionName) {
 const BaseProvider = class BaseProvider {
     constructor() {
         this.logger = logger;
-        this.userAgent = randomUseragent.getRandom();
+        this.userAgent = useragent.getUserAgent();
 
         if (new.target === BaseProvider) {
             throw new TypeError("Cannot construct BaseProvider instances directly");
@@ -195,6 +196,55 @@ const BaseProvider = class BaseProvider {
             }
         }
         this.logger.error(`${this.getProviderId()}: An unexpected error occurred:`, e);
+    }
+
+    /**
+     *
+     * @param {Object} $ Cheerio object.
+     * @param {Function} processFn Optional callback for modifying the default jQuery object.
+     * @return {Function}
+     */
+    _getJqueryShim($, processFn = null) {
+        return javascriptEval._getJqueryShim($, processFn);
+    }
+
+    /**
+     * Return a JWPlayer shim object.
+     * @param {Function} setupCallback
+     */
+    _getJwPlayerShim(setupCallback) {
+        return javascriptEval._getJwPlayerShim(setupCallback);
+    }
+
+    _getClapprShims(playerCallback) {
+        return javascriptEval._getClapprShims(playerCallback);
+    }
+
+    _getDefaultSandbox(jQuery, jwPlayer, clapper, includeBrowserShims) {
+        return javascriptEval._getDefaultSandbox(jQuery, jwPlayer, clapper, includeBrowserShims);
+    }
+
+    /**
+     * Create a native shim object e.g. Document, which allows you call any functions/properties and allow them
+     * to resolve without throwing a "Cannot read property of undefined" error.
+     * Useful for bypassing tricky resolvers like Openload which check for browser objects.
+     *
+     * @param {String} propertyName
+     * @param {Boolean} useProxy Set this if you need to shim complex objects with dynamic properties.
+     * @return {Function|Proxy}
+     */
+    _createNativeProxyShim(propertyName, useProxy) {
+        return javascriptEval._createNativeProxyShim(propertyName, useProxy);
+    }
+
+    /**
+     * Resolve jwplayer links.
+     * @param {Object} setupConfig
+     * @param {Object} meta
+     * @return {ClawsLink[]}
+     */
+    _resolveJwPlayerLinks(setupConfig, meta) {
+        return javascriptEval._resolveJwPlayerLinks(setupConfig, meta);
     }
 };
