+++ src/scrapers/resolvers/_BaseResolver.js
@@ -1,7 +1,8 @@
 const createEvent = require('../../utils/createEvent');
 const logger = require('../../utils/logger');
+const JavascriptEval = require('../../service/javascriptEval');
+const useragent = require('../../service/useragent');
 const rp = require('request-promise');
-const randomUseragent = require('random-useragent');
 
 function _implementMe(functionName) {
     throw new Error(`Must implement ${functionName}()`);
@@ -76,11 +77,7 @@ const BaseResolver = class BaseResolver {
      * @return String
      */
     getDefaultUserAgent() {
-        return process['CLAWS_DEFAULT_UA'] || randomUseragent.getRandom((ua) => {
-            // don't include questionable user-agents.
-            let excludeGroups = /Spider|Legacy|Console|Miscellaneous/;
-            return !excludeGroups.test(ua.folder);
-        });
+        return useragent.getUserAgent();
     }
 
     /**
@@ -253,26 +250,7 @@ const BaseResolver = class BaseResolver {
      * @return {Function}
      */
     _getJqueryShim($, processFn = null) {
-        return function (selector, anotherArg) {
-
-            let jQuery = function () {
-                return jQuery;
-            };
-
-            jQuery.ready = jQuery.click = jQuery.hide = jQuery.show = jQuery.mouseup = jQuery.mousedown =
-                jQuery.hasClass = jQuery.attr = jQuery.post = jQuery.get = jQuery.append = jQuery;
-
-            jQuery.cookie = function () {
-                return true;
-            };
-
-            if (processFn) {
-                // Allow sub-classes to modify the returned jQuery object.
-                processFn(jQuery, $, arguments);
-            }
-
-            return jQuery
-        };
+        return JavascriptEval._getJqueryShim($, processFn);
     }
 
     /**
@@ -280,86 +258,15 @@ const BaseResolver = class BaseResolver {
      * @param {Function} setupCallback
      */
     _getJwPlayerShim(setupCallback) {
-        let jwplayer = function () {
-            return jwplayer;
-        };
-        jwplayer.on = jwplayer.addButton = jwplayer.onTime = jwplayer.onComplete = jwplayer;
-
-        jwplayer.setup = setupCallback;
-
-        return jwplayer;
+        return JavascriptEval._getJwPlayerShim(setupCallback);
     }
 
     _getClapprShims(playerCallback) {
-        let noop = function () {
-            return noop;
-        };
-        let player = function (config) {
-            playerCallback && playerCallback(config, 'constructor');
-            return {
-                options: config,
-                attachTo: noop,
-                listenTo: noop,
-                configure: noop,
-                play: noop,
-                load: (options) => {
-                    playerCallback && playerCallback(options, 'load');
-                },
-            }
-        };
-        let events = function () {
-            return {
-                PLAYER_FULLSCREEN: 'fullscreen',
-                on: events,
-                once: events,
-                off: events,
-                trigger: events,
-                stopListening: events,
-            }
-        };
-        let Clappr = {
-            Player: player,
-            Events: events,
-        };
-        Clappr.$ = function () {
-            return {text: noop};
-        };
-
-        // List of properties which should be merged into the context object.
-        return {
-            Clappr: Clappr,
-            LevelSelector: {},
-            TheaterMode: {},
-        };
+        return JavascriptEval._getClapprShims(playerCallback);
     }
 
     _getDefaultSandbox(jQuery, jwPlayer, clapper, includeBrowserShims) {
-        let sandbox = {
-            $: jQuery,
-            jQuery: jQuery,
-            jwplayer: jwPlayer,
-            sin: Math.sin,
-            navigator: {
-                userAgent: ''
-            },
-            atob: function (encodedData) {
-                return Buffer.from(encodedData, 'base64').toString();
-            },
-            btoa: function (stringToEncode) {
-                return Buffer.from(stringToEncode).toString('base64');
-            },
-            ...clapper,
-        };
-
-        if (includeBrowserShims) {
-            let Document = this._createNativeProxyShim('Document', false);
-
-            sandbox['Document'] = Document;
-            sandbox['document'] = Document;
-        }
-
-        sandbox['window'] = sandbox;
-        return sandbox;
+        return JavascriptEval._getDefaultSandbox(jQuery, jwPlayer, clapper, includeBrowserShims);
     }
 
     /**
@@ -372,53 +279,7 @@ const BaseResolver = class BaseResolver {
      * @return {Function|Proxy}
      */
     _createNativeProxyShim(propertyName, useProxy) {
-        let that = this;
-
-        let shim;
-        if (useProxy) {
-            shim = new Proxy(function () {
-                return shim;
-            }, {
-                get: function (target, name, proxy) {
-                    if (target.hasOwnProperty(name)) {
-                        return target[name];
-                    }
-
-                    if ('hasOwnProperty' === name) {
-                        return function () {
-                            return true;
-                        }
-                    } else if ('toString' === name) {
-                        return `toString: ${name}`;
-                    } else if ('length' === name) {
-                        return 0;
-                    } else if (name === '__proto__') {
-                        return proxy ? proxy : that;
-                    }
-
-                    return that._createNativeProxyShim(name, true);
-                },
-            });
-        } else {
-            shim = function () {
-            };
-        }
-
-        let objectToString = "[object " + (propertyName.charAt(0).toUpperCase() + propertyName.slice(1)) + "]";
-
-        shim.toString = function () {
-            if (useProxy || /^[A-Z]/.test(propertyName)) {
-                // Starts with a uppercase - it's a class.
-                return "function " + propertyName + "() { [native code] }";
-            }
-            return objectToString;
-        };
-        shim.__proto__.toString = shim.toString;
-        shim.prototype.toString = function () {
-            return objectToString
-        };
-
-        return shim;
+        return JavascriptEval._createNativeProxyShim(propertyName, useProxy);
     }
 
     /**
@@ -428,26 +289,7 @@ const BaseResolver = class BaseResolver {
      * @return {ClawsLink[]}
      */
     _resolveJwPlayerLinks(setupConfig, meta) {
-        let links = [];
-        if (setupConfig && setupConfig.file) {
-            links.push({
-                data: setupConfig.file,
-                meta,
-            });
-        }
-        if (setupConfig.sources) {
-            setupConfig.sources.forEach(function (source) {
-                links.push({
-                    data: source.file,
-                    meta: {
-                        ...meta,
-                        quality: source.label,
-                    },
-                });
-            });
-        }
-
-        return links;
+        return JavascriptEval._resolveJwPlayerLinks(setupConfig, meta);
     }
 };
